Refactorizar comentarios
Ahora Commentable se encarga de obtener el contenido del comentario. El
componente Reply ya no es necesario.
