--- src/comments/components/Reply.js
@@ -1,11 +1,0 @@
-import React, { PureComponent } from 'react'
-
-import Commentable from './Commentable'
-
-class Reply extends PureComponent {
-  render() {
-    return <Commentable {...this.props} />
-  }
-}
-
-export default Reply
